app/pyproject.toml
[project]
name = "new-dir"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs>=2.6.1",
    "aiohttp>=3.12.15",
    "aiosignal>=1.4.0",
    "attrs>=25.3.0",
    "blinker>=1.9.0",
    "certifi>=2025.8.3",
    "charset-normalizer>=3.4.3",
    "click>=8.2.1",
    "colorama>=0.4.6",
    "flask>=3.1.1",
    "flask-httpauth>=4.8.0",
    "frozenlist>=1.7.0",
    "greenlet>=3.2.4",
    "idna>=3.10",
    "iniconfig>=2.1.0",
    "markupsafe>=3.0.2",
    "mccabe>=0.7.0",
    "multidict>=6.6.4",
    "packaging>=25.0",
    "pluggy>=1.6.0",
    "psycopg2>=2.9.10",
    "pycodestyle>=2.14.0",
    "pydantic>=2.11.7",
    "pyflakes>=3.4.0",
    "pygments>=2.19.2",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.43",
    "typing-extensions>=4.14.1",
    "urllib3>=2.5.0",
    "werkzeug>=3.1.3",
    "wheel>=0.45.1",
    "xlsxwriter>=3.2.5",
    "yarl>=1.20.1",
]
